requirements.txt
alembic==1.8.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Flask==2.1.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
instaloader==4.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
pkg_resources==0.0.0
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.11
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1
